Seed class extracted
