+++ thread/Sprites/Seed.swift
@@ -8,13 +8,12 @@
 import Foundation
 import SpriteKit
 
-extension GameScene {
+class Seed {
 
    func createSeed() -> SKSpriteNode {
     // Assign image and position
     let seed = SKSpriteNode(imageNamed: "threadship")
     seed.size = CGSize(width: 20, height: 20)
-    seed.position = CGPoint(x: self.frame.midX, y: self.frame.midY)
     
     seed.physicsBody = SKPhysicsBody(texture: seed.texture!, size: seed.size)
    
@@ -34,35 +33,12 @@ extension GameScene {
     return seed
       
   }
+}
 
     
 // MOTION
 
-func seedThrust() {
-
-    seed.physicsBody?.applyImpulse(CGVector(dx: 0, dy: 0.4))
-    seed.physicsBody?.applyTorque(CGFloat(0.01))
-  //  seed.physicsBody?.velocity = CGVector(dx: 0, dy: 90)
-}
-
     
-  func seedFlight() {
-    if isLeft == true {
-        seed.physicsBody?.applyForce(CGVector(dx: -0.3, dy: 0.2))
-       // seed.physicsBody?.velocity = CGVector(dx: -50, dy: 50)
-
-    }
-    else if isRight == true {
-        seed.physicsBody?.applyForce(CGVector(dx: 0.3, dy: 0.2))
-
-        //seed.physicsBody?.velocity = CGVector(dx: 50, dy: 50)
-
-    }
-
-    
-}
-    
-}
 
 // PHYSICS NOTES
